@@ -3,18 +3,18 @@ package ru.yandex.practicum.filmorate.model;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-import java.time.LocalDate;
+import java.sql.Date;
 
-public class MinimumDateValidator implements ConstraintValidator<MinimumDate, LocalDate> {
-    private LocalDate minimumDate;
+public class MinimumDateValidator implements ConstraintValidator<MinimumDate, Date> {
+    private Date minimumDate;
 
     @Override
     public void initialize(MinimumDate constraintAnnotation) {
-        minimumDate = LocalDate.parse(constraintAnnotation.value());
+        minimumDate = Date.valueOf(constraintAnnotation.value());
     }
 
     @Override
-    public boolean isValid(LocalDate value, ConstraintValidatorContext context) {
-        return value == null || !value.isBefore(minimumDate);
+    public boolean isValid(Date value, ConstraintValidatorContext constraintValidatorContext) {
+        return value != null && !value.before(minimumDate);
     }
 }
