@@ -12,8 +12,7 @@ import java.util.List;
 import java.util.Map;
 
 @Slf4j
-@Repository("inMemoryUserStorage")
-@Primary
+@Repository
 public class InMemoryUserStorage implements UserStorage {
 
     private final Map<Integer, User> users = new HashMap<>();
